chore: update templates
