+++ tools/kystrap/templates/go/.dockerignore
@@ -1,4 +1,5 @@
 **/out
+build/
 .git
 .env
 *.md
