+++ tools/kystrap/templates/go/.gitignore
@@ -13,6 +13,7 @@
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+build/
 
 # Dependency directories (remove the comment below to include it)
 # vendor/
